backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
requests>=2.31,<3.0
pydantic>=2.0,<3.0
pywebview>=6,<7
